+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Changed
+- metrics-cassandra-graphite.rb: Added support for Row Cache in Cassandra 3.0+
 
 ## [2.1.0] - 2019-03-04
 ### Added